Original version of songsbro
There are chances that the source code might reveal a lot of personal information or personally identifiable information about me. But I am uploading this just before destroying the site, so got no time to rethink. Please feel free to adopt this code or fork it and build something genius out of it.